chore: move syllab adaptor to project root and create examples dir
