+++ adapter/scylladb/db.go
