+++ adapter/scylladb/sessionx.go
